Move throttler to its own package
