+++ sources/dynamodb/stream.go
@@ -4,37 +4,15 @@ import (
 	"context"
 	"fmt"
 	"log/slog"
-	"sync"
 	"time"
 
 	"github.com/artie-labs/reader/config"
+	"github.com/artie-labs/reader/lib/throttler"
 	"github.com/artie-labs/reader/sources/dynamodb/offsets"
 	"github.com/artie-labs/reader/writers"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/dynamodbstreams"
 )
-
-type Throttler struct {
-	limit   int64
-	running int64
-	mu      sync.Mutex
-}
-
-func (t *Throttler) Start() {
-	t.mu.Lock()
-	t.running++
-	t.mu.Unlock()
-}
-
-func (t *Throttler) Done() {
-	t.mu.Lock()
-	t.running--
-	t.mu.Unlock()
-}
-
-func (t *Throttler) Allowed() bool {
-	return t.running < t.limit
-}
 
 type StreamStore struct {
 	tableName string
@@ -44,7 +22,7 @@ type StreamStore struct {
 	streams   *dynamodbstreams.DynamoDBStreams
 	storage   *offsets.OffsetStorage
 	shardChan chan *dynamodbstreams.Shard
-	throttler *Throttler
+	throttler *throttler.Throttler
 }
 
 func (s *StreamStore) Close() error {
